Refactor a ton of this garbage :|
Having everything in one line is not a good solution. Split off all of
the pico-8 api stuff into api.lua, and all of the cartridge loading
stuff into cart.lua, as much as possible.

There are likely still regressions present here but most carts are back
to being up and running.