Clean up audio and add Android overlay
Audio is now generated when the source says it has buffers available.
Mix the four channels together ourselves to prevent desync.

Adjust the volumes of the waveforms
Fix the detuned triangle waveform
Quiet noise a bit, it's volume is still not accurate but it's better.
Clean up the audio timing with a measured number.
Fix the way the pattern timing is sleected.

Add in an Android overlay, and touchscreen support
This might not be drawing correctly ... but I'd need more phones for
testing this out.

Accidentally removed printh, oops.